requirements.txt
opencv-python
numpy
pandas
matplotlib
py-feat
